+++ packages/component/src/block/orderEntry/index.tsx
@@ -304,7 +304,7 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
                   value: OrderSide.BUY,
                   disabled,
                   activeClassName:
-                    "orderly-bg-success-light orderly-text-base-contrast after:orderly-bg-success-light",
+                    "orderly-bg-trade-profit orderly-text-base-contrast after:orderly-bg-trade-profit",
                   disabledClassName:
                     "orderly-bg-base-400 orderly-text-base-contrast-20 after:orderly-bg-base-400 orderly-cursor-not-allowed",
                 },
@@ -313,7 +313,7 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
                   value: OrderSide.SELL,
                   disabled,
                   activeClassName:
-                    "orderly-bg-danger-light orderly-text-base-contrast after:orderly-bg-danger-light",
+                    "orderly-bg-trade-loss orderly-text-base-contrast after:orderly-bg-trade-loss",
                   disabledClassName:
                     "orderly-bg-base-400 orderly-text-base-contrast-20 after:orderly-bg-base-400 orderly-cursor-not-allowed",
                 },
@@ -340,7 +340,7 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
                 size={"small"}
                 type="button"
                 onClick={onDeposit}
-                className="orderly-text-primary orderly-text-4xs"
+                className="orderly-text-link orderly-text-4xs"
               >
                 Deposit
               </Button>
@@ -354,7 +354,7 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
                   <MSelect
                     label={"Order Type"}
                     value={field.value}
-                    className="orderly-bg-base-600"
+                    className="orderly-bg-base-600 orderly-font-semibold"
                     color={side === OrderSide.BUY ? "buy" : "sell"}
                     fullWidth
                     options={[
@@ -403,7 +403,7 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
                     // placeholder={"Market"}
                     helpText={methods.formState.errors?.order_price?.message}
                     value={isMarketOrder ? "Market" : field.value}
-                    className="orderly-text-right"
+                    className="orderly-text-right orderly-font-semibold"
                     containerClassName={
                       isMarketOrder
                         ? "orderly-bg-base-700"