+++ packages/component/src/block/walletConnect/walletConnect.tsx
@@ -96,7 +96,7 @@ export const WalletConnect: FC<WalletConnectProps> = (props) => {
         Signing is free.
       </div>
 
-      <Paper className="orderly-bg-base-300">
+      <Paper className="orderly-bg-base-500">
         <ListTile
           className="orderly-text-xs desktop:orderly-text-base"
           avatar={
@@ -132,11 +132,8 @@ export const WalletConnect: FC<WalletConnectProps> = (props) => {
         />
       </Paper>
 
-      <div className="orderly-pt-5 orderly-pb-7 orderly-flex orderly-justify-between orderly-items-center orderly-">
-        <div
-          className="orderly-text-base-contrast-54 orderly-text-xs desktop:orderly-text-base"
-          onClick={showRememberHint}
-        >
+      <div className="orderly-pt-5 orderly-pb-7 orderly-flex orderly-justify-between orderly-items-center">
+        <div className="orderly-text-base-contrast-54 orderly-text-xs  desktop:orderly-text-base" onClick={showRememberHint}>
           <span>Remember me</span>
           <InfoIcon className="orderly-inline-block orderly-ml-2" size={14} />
         </div>