+++ packages/core/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @orderly.network/core
 
+## 0.1.19
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/types@0.1.19
+
+## 0.1.18
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/types@0.1.18
+
+## 0.1.17
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/types@0.1.17
+
+## 0.1.16
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/types@0.1.16
+
+## 0.1.15
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/types@0.1.15
+
 ## 0.1.14
 
 ### Patch Changes