+++ packages/net/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @orderly.network/net
 
+## 1.0.106
+
+### Patch Changes
+
+- fix bug
+
+## 1.0.105
+
+### Patch Changes
+
+- fix bug
+
+## 1.0.104
+
+### Patch Changes
+
+- fix bug
+
+## 1.0.103
+
+### Patch Changes
+
+- fix bug
+
+## 1.0.102
+
+### Patch Changes
+
+- fix bug
+
 ## 1.0.101
 
 ### Patch Changes